Stop homing after float value
The missile will now stop homing on two occasions: when it has come too close to the player and when it has been following for a specific amount of time. These can both be edited in the homing projectile prefab. Do note that changing the movement speed will most likely require tweaking of these values.